requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==5.0
django-admin-sortable2==2.2.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
pillow==11.0.0
sqlparse==0.5.1
tornado==6.4.1
whitenoise==6.8.2
